+++ CHANGELOG.md
@@ -2,5 +2,7 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Added
+- Protocol for unmanaged resources, defining a method to free the resource
 
 [Unreleased]: https://github.com/IGJoshua/s-expresso/compare/master..develop